Add HTTP Sender
This will use the Jaeger collector HTTP endpoint (assumed to be running
on localhost) by default, but is configurable to send Thrift-encoded
spans to any HTTP endpoint.

Spans are batched upto a certain amount (based on # of spans and not
byte lengths) and are also flushed on a regular interval by the
RemoteReporter just like with the UDPSender.

It also supports basic auth in case the Jaeger collector has some form
of authentication in front of it.

Fixes #170

Signed-off-by: Ben Keith <bkeith@signalfx.com>