Change Zipkin Codec to Not Inject Missing Parent
The Zipkin b3 TextMap codec was injecting an empty string for the
x-b3-parentspanid field when one was not present in the object it
extracted the SpanContext from. This caused gaps in the traces.

This change first checks the span context's parentIdStr value,
and does not inject it into the carrier if it is falsy.

Signed-off-by: Andrew DiNunzio <adinunzio84@gmail.com>