Allow jaeger-client to be bundled
Always reading the jaeger thrift definition file from the root dir (src)
allows consumers to bundle jaeger-client, e.g. for uploading code to an
AWS lambda function.

To bundle jaeger-client with webpack for example, the thrift file needs
to be copied into the output directory of the bundle like this:

```js
{
  plugins: [
    new CopyPlugin([
      {
        from: require.resolve(
          'jaeger-client/dist/src/jaeger-idl/thrift/jaeger.thrift'
        ),
        to: 'jaeger-idl/thrift/jaeger.thrift'
      }
    ])
  ]
}
```

With this in place `fs.readFileSync` will be able to find the file.

Signed-off-by: Hendrik Liebau <mail@hendrik-liebau.de>