Simplify bundling udp_sender (#441, #530)
Add a helper function in src/thrift.js that returns an absolute path to
agent.thrift, without using "../". This is done to simplify bundling with
webpack or esbuild: assume an app is bundled to /app/index.js, if we resolve
a path to "../thrift-idl/etc", the file will be loaded from /thrift-idl/etc,
which is at the root of the filesystem.

Adjust README to support udp-sender.

This commit is a continuation of #441.

Signed-off-by: Thorsten Nadel <18720251+punkfest@users.noreply.github.com>