Bind to localhost instead of INADDR_ANY for local reporter UDP socket creation
Before this change, we always bound UDP to all interfaces (::0 and 0.0.0.0) to
create an ephemeral port for reporting. Binding to INADDR_ANY for UDP ports
increases the attack surface for potential denial of service since INADDR_ANY
UDP ports will always accept all traffic that comes its way.

The proposed fix is to only bind to localhost for both ipv4 and ipv6, when the
configured HOST is localhost. For other hosts, we can't easily tell if we have
to bind to INADDR_ANY.

I understand that since we're using an ephemeral port the risk isn't quite as
big for DoS as a privileged port, however we're doing some network cleanup and
found that a few of our applications unnecessarily use INADDR_ANY, so making
this fix would clean up a lot of these patterns.

Signed-off-by: Thorsten Nadel <18720251+punkfest@users.noreply.github.com>